tab menu like branch of  novel game
Please review and comment it to make this code simpler or rewrite with javascript.
I wrote a code to make my blog interesting. 
Using this code,  blog visitor can choice a story to read, from some choices.
But I don't have so much experience to write codes, so please check if it is correct or not.